feat: HTTPSession interface to wrap and inject http library
