+++ pyproject.toml
@@ -31,8 +31,9 @@ dependencies = ["aiohttp~=3.7"]
 
 [project.optional-dependencies]
 tests = [
-    "pytest>=6", # pyproject.toml support
     "coverage[toml]>=5", # pyproject.toml support
+    "pytest>=6", # pyproject.toml support
+    "pytest-asyncio",
 ]
 tools = [
     "hatch",
@@ -44,6 +45,9 @@ dev = ["shmdash[tests,tools]"] # recursive dependency since pip 21.2
 [project.urls]
 Source = "https://github.com/vallen-systems/pySHMdash"
 Issues = "https://github.com/vallen-systems/pySHMdash/issues"
+
+[tool.hatch.envs.hatch-test]
+features = ["tests"]
 
 [[tool.hatch.envs.hatch-test.matrix]]
 python = ["3.8", "3.9", "3.10", "3.11", "3.12"]
@@ -86,6 +90,8 @@ addopts = "-ra -q" # test summary for (a)ll except passed
 testpaths = ["tests"]
 log_cli = true
 log_cli_level = "WARNING"
+asyncio_mode = "auto"
+asyncio_default_fixture_loop_scope = "module"
 
 [tool.coverage.run]
 branch = true
